feat: Update package.json version to 2.0.7
This commit updates the version number in the package.json file to 2.0.7. This change enables

feat: Update package.json version to 2.0.7

This commit updates the version number in package.json to 2.0.7. The new version ensures that the latest changes are properly identified. No functional changes are introduced as a result of this update.

Commit message generated with the help of Assistant